Fix ObjectCache expiring newly referenced object from the DatabasePager thread.
This fixing a problem that results in a "Deleting still referenced object".

When the DatabasePager has just referenced something that is in the object cache after releasing the objectCacheMutex when the ObjectCache expiry method was waiting for this mutex - the reference happens immediately after the mutex is released in the pager thread; however the object meets the expiry time condition because it had previously expired. So there exists a condition where the object may be erased from the cache (and deleted) when it has just been referenced and this will often lead to a segfault.

This fix uses a ReadWriteMutex to control the expiry process. The database pager is considered to be a reader, of which there can be many, and only when there are no readers will the object cache expiry (considered a writer) will be able to remove items from the cache. This will defers the expiry of objects to a later frame.

Possible side effects of this fix could be an unusually large number of items being expired when the mutex is released; however in testing the longest was 743 frames and generally <10 frames.

(cherry picked from commit cc6d41a1195a8e4ce1e01dd58a32ad42f212a079)

(cherry picked from commit 568c820a3607aba24844cf7b73a0d639c3933168)